+++ packages/react-components/react-datepicker/src/components/CalendarDayGrid/index.ts
@@ -1,0 +1,5 @@
+export * from './CalendarDayGrid';
+export * from './CalendarDayGrid.types';
+// export * from './renderCalendarDayGrid';
+// export * from './useCalendarDayGrid';
+export * from './useCalendarDayGridStyles';