+++ packages/react-components/react-datepicker/src/components/CalendarYear/index.ts
@@ -1,5 +1,5 @@
 export * from './CalendarYear';
 export * from './CalendarYear.types';
-export * from './renderCalendarYear';
-export * from './useCalendarYear';
+// export * from './renderCalendarYear';
+// export * from './useCalendarYear';
 export * from './useCalendarYearStyles';