+++ packages/react-components/react-provider/src/components/FluentProvider/FluentProvider.types.ts
@@ -1,4 +1,5 @@
 import type {
+  OverridesContextValue_unstable as OverridesContextValue,
   ProviderContextValue_unstable as ProviderContextValue,
   TooltipVisibilityContextValue_unstable as TooltipVisibilityContextValue,
   ThemeClassNameContextValue_unstable as ThemeClassNameContextValue,
@@ -26,16 +27,19 @@ export type FluentProviderProps = Omit<ComponentProps<FluentProviderSlots>, 'dir
 
   /** Sets the theme used in a scope. */
   theme?: PartialTheme;
+
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  overrides_unstable?: OverridesContextValue;
 };
 
 export type FluentProviderState = ComponentState<FluentProviderSlots> &
   Pick<FluentProviderProps, 'targetDocument'> &
-  Required<Pick<FluentProviderProps, 'applyStylesToPortals' | 'dir'>> & {
+  Required<Pick<FluentProviderProps, 'applyStylesToPortals' | 'dir' | 'overrides_unstable'>> & {
     theme: ThemeContextValue;
     themeClassName: string;
   };
 
-export type FluentProviderContextValues = Pick<FluentProviderState, 'theme'> & {
+export type FluentProviderContextValues = Pick<FluentProviderState, 'theme' | 'overrides_unstable'> & {
   provider: ProviderContextValue;
   themeClassName: ThemeClassNameContextValue;
   textDirection: 'ltr' | 'rtl';