+++ packages/react-components/react-tree/src/components/TreeItem/renderTreeItem.tsx
@@ -1,23 +1,24 @@
 import * as React from 'react';
 import { getSlots } from '@fluentui/react-utilities';
-import type { TreeItemState, TreeItemSlots } from './TreeItem.types';
+import type { TreeItemState, TreeItemSlots, TreeItemContextValues } from './TreeItem.types';
+import { TreeItemProvider } from '../../contexts/treeItemContext';
 
 /**
  * Render the final JSX of TreeItem
  */
-export const renderTreeItem_unstable = (state: TreeItemState) => {
+export const renderTreeItem_unstable = (state: TreeItemState, contextValues: TreeItemContextValues) => {
   const { slots, slotProps } = getSlots<TreeItemSlots>(state);
 
   return (
-    <slots.groupper {...slotProps.groupper}>
+    <TreeItemProvider value={contextValues.treeItem}>
       <slots.root {...slotProps.root}>
-        {slots.expandIcon && <slots.expandIcon {...slotProps.expandIcon} />}
-        {slots.iconBefore && <slots.iconBefore {...slotProps.iconBefore} />}
-        {slotProps.root.children}
-        {slots.iconAfter && <slots.iconAfter {...slotProps.iconAfter} />}
-        {slots.badges && <slots.badges {...slotProps.badges} />}
-        {slots.actions && <slots.actions {...slotProps.actions} />}
+        <slots.content {...slotProps.content}>
+          {slots.expandIcon && <slots.expandIcon {...slotProps.expandIcon} />}
+          {slotProps.content.children}
+          {slots.actions && <slots.actions {...slotProps.actions} />}
+        </slots.content>
+        {slots.subtree && <slots.subtree {...slotProps.subtree} />}
       </slots.root>
-    </slots.groupper>
+    </TreeItemProvider>
   );
 };