+++ packages/react-components/react-tree/stories/TreeItemPersonaLayout/index.stories.tsx
@@ -1,0 +1,17 @@
+import { TreeItemPersonaLayout } from '@fluentui/react-tree';
+
+import descriptionMd from './TreeItemPersonaLayoutDescription.md';
+
+export { Default } from './TreeItemPersonaLayoutDefault.stories';
+
+export default {
+  title: 'Preview Components/Tree/PersonaTreeItem',
+  component: TreeItemPersonaLayout,
+  parameters: {
+    docs: {
+      description: {
+        component: [descriptionMd].join('\n'),
+      },
+    },
+  },
+};